rename: locationmodalprovider 폴더이동(#306)
